doc: add limitation in using segment-sz in crypto-perf
When segment-sz is given as parameter,
use buffer-sz in descending order.
The reason is, same scatter gather buffer list is being used
for all packet sizes which is initialized once in the beginning.
Now if a small packet case is run first,
mbuf SG list will be trimmed to a smaller number of SG and
rest of the SGs are lost resulting in segfault.
Buffers cannot be re-initialized as that will come in datapath and
will be costly.
So to workaround this, if segment-sz is passed as argument,
use values in buffer-sz in descending order.

Signed-off-by: Akhil Goyal <gakhil@marvell.com>
Acked-by: Anoob Joseph <anoobj@marvell.com>
Acked-by: Hemant Agrawal <hemant.agrawal@nxp.com>